requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
blis==0.7.11
CacheControl==0.14.3
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.16.0
confection==0.1.5
cryptography==45.0.5
cymem==2.0.11
diff-match-patch==20241021
dj-database-url==3.0.1
Django==4.2.21
django-cors-headers==4.7.0
django-filter==25.1
django-import-export==4.3.8
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.7.0/es_core_news_sm-3.7.0-py3-none-any.whl
exceptiongroup==1.3.0
filelock==3.18.0
firebase_admin==7.1.0
fsspec==2025.5.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.3
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.4
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
MarkupSafe==3.0.2
mongoengine==0.29.1
mpmath==1.3.0
msgpack==1.1.1
murmurhash==1.0.13
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu12==12.1.105
packaging==25.0
pillow==11.2.1
preshed==3.0.10
proto-plus==1.26.1
protobuf==6.31.1
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.12.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
safetensors==0.5.3
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
sympy==1.14.0
tablib==3.8.0
thinc==8.2.5
tokenizers==0.19.1
torch==2.2.2
tqdm==4.67.1
transformers==4.40.0
triton==2.2.0
typer==0.9.4
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
wasabi==1.1.3
weasel==0.3.4
